Refactor:Refactor project to subdirectories in source folder
